# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Order Processing System API
  description: Backend API for creating, tracking, and managing e-commerce orders.
  version: 1.0.0
  contact:
    name: Tushar Sharma
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local

tags:
  - name: Orders
    description: Create, track, and manage e-commerce orders

paths:
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponse' }
        '400':
          description: Validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - { name: status, in: query, required: false, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - { name: customerId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: page, in: query, required: false, schema: { type: integer, default: 0 } }
        - { name: size, in: query, required: false, schema: { type: integer, default: 20 } }
        - { name: sortBy, in: query, required: false, schema: { type: string, default: createdAt } }
        - { name: sortDir, in: query, required: false, schema: { type: string, default: desc } }
      responses:
        '200':
          description: Paged list of orders
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedOrderResponse' }

  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order by id
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Order found, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponse' } } } }
        '404': { description: Order not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }

  /api/v1/orders/number/{orderNumber}:
    get:
      tags: [Orders]
      summary: Get an order by its order number
      parameters:
        - { name: orderNumber, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Order found, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponse' } } } }
        '404': { description: Order not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }

  /api/v1/orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrderStatusRequest' }
      responses:
        '200': { description: Status updated, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponse' } } } }
        '404': { description: Order not found }
        '409': { description: Illegal status transition }

  /api/v1/orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order (only if PENDING)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Order cancelled, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponse' } } } }
        '404': { description: Order not found }
        '409': { description: Order is not PENDING }

  /api/v1/orders/statistics:
    get:
      tags: [Orders]
      summary: Order statistics
      responses:
        '200':
          description: Aggregate statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderStatisticsResponse' }

components:
  schemas:
    OrderStatus:
      type: string
      enum: [PENDING, PROCESSING, SHIPPED, DELIVERED, CANCELLED]

    CreateOrderItemRequest:
      type: object
      required: [productId, productSku, productName, quantity, unitPrice]
      properties:
        productId: { type: integer, format: int64 }
        productSku: { type: string }
        productName: { type: string }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, format: double }

    CreateOrderRequest:
      type: object
      required: [customerId, items]
      properties:
        customerId: { type: integer, format: int64 }
        customerEmail: { type: string, format: email }
        customerName: { type: string }
        shippingAddress: { type: string }
        notes: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/CreateOrderItemRequest' }

    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/OrderStatus' }
        reason: { type: string }

    OrderItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        productId: { type: integer, format: int64 }
        productSku: { type: string }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number, format: double }
        lineTotal: { type: number, format: double }

    OrderResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderNumber: { type: string }
        customerId: { type: integer, format: int64 }
        customerEmail: { type: string }
        customerName: { type: string }
        shippingAddress: { type: string }
        notes: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        totalAmount: { type: number, format: double }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItemResponse' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    PagedOrderResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/OrderResponse' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }

    OrderStatisticsResponse:
      type: object
      properties:
        totalOrders: { type: integer, format: int64 }
        countByStatus:
          type: object
          additionalProperties: { type: integer, format: int64 }
        totalRevenue: { type: number, format: double }

    ApiError:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }